% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_area.R
\name{rh_area}
\alias{rh_area}
\title{Calculate Redlined Area per Tract}
\usage{
rh_area(areal_unit, holc, cat)
}
\arguments{
\item{areal_unit}{\code{sf} object with census GEOID column}

\item{holc}{\code{sf} object with redlining geometry}

\item{cat}{string input that is one of \code{"A"}, \code{"B"}, \code{"C"}, or \code{"D"} or
a vector like \code{c("A", "B")}}
}
\value{
A tibble with a row for each tract GEOID and the corresponding area redlined for a given category.
}
\description{
This function takes a census tract input as well as a HOLC redlining input and
   calculates the areas per census tract that have been redlined for a given category.
}
