% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_area_join.R
\name{rh_area_join}
\alias{rh_area_join}
\title{Combined Measured Areas with Tract Boundaries}
\usage{
rh_area_join(areal_unit, area, by, cat)
}
\arguments{
\item{areal_unit}{\code{sf} object with census GEOID column}

\item{area}{A tbl or data frame containing area of redlining per tract for a given grade}

\item{by}{Variable name common to both \code{tract} and \code{area} (such as \code{GEOID})}

\item{cat}{Category name to be given to new variable in the returned \code{sf} object}
}
\value{
An \code{sf} object with a column containing the area covered by a given grade.
}
\description{
This function takes a census tract output as well as the returned results of
    \code{rh_area} and combines them to provide area measurements for each feature regardless
    of if it was redlined or not.
}
