% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stl_holc_1937}
\alias{stl_holc_1937}
\title{St. Louis Home Owners Loan Corporation Boundaries, 1937}
\format{
A data frame with 126 rows and 3 variables:
\describe{
  \item{holc_id}{feature id}
  \item{holc_grade}{one of A, B, C, or D}
  \item{geometry}{simple features geometry}
  }
}
\usage{
data(stl_holc_1937)
}
\description{
A simple features data set originally created by Colin Gordon, Ph.D. for
    his Mapping Decline project. See \url{https://worldmap.harvard.edu/mappingdecline/}.
}
\keyword{datasets}
